+++ src/main/java/price_algorithm/PriceAlgorithmConstant.java
@@ -4,7 +4,7 @@ import static parking_manager_agent.util.Constants.BASE_PRICE;
 
 public class PriceAlgorithmConstant implements PriceAlgorithm {
     @Override
-    public double calculatePrice(double numOfOccupiedPlaces, double capacity) {
+    public double calculatePrice(int numOfOccupiedPlaces, int capacity) {
         return BASE_PRICE;
     }
 }
